MINOR-restructuring: moved PartitionMetadataClient and NetworkPartitionMetadataClient
