+++ server/src/test/java/org/apache/kafka/server/NetworkPartitionMetadataClientTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.coordinator.group;
+package org.apache.kafka.server;
 
 import org.apache.kafka.clients.ClientResponse;
 import org.apache.kafka.clients.CommonClientConfigs;
@@ -177,14 +177,14 @@ class NetworkPartitionMetadataClientTest {
         Set<TopicPartition> partitions = new HashSet<>();
         partitions.add(tp);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
         assertEquals(1, futures.size());
         assertTrue(futures.containsKey(tp));
 
-        PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
         assertTrue(futures.get(tp).isDone() && !futures.get(tp).isCompletedExceptionally());
         assertNotNull(response);
         assertEquals(Errors.NONE.code(), response.error().code());
@@ -207,14 +207,14 @@ class NetworkPartitionMetadataClientTest {
         Set<TopicPartition> partitions = new HashSet<>();
         partitions.add(tp);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
         assertEquals(1, futures.size());
         assertTrue(futures.containsKey(tp));
 
-        PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
         assertTrue(futures.get(tp).isDone() && !futures.get(tp).isCompletedExceptionally());
         assertNotNull(response);
         assertEquals(Errors.LEADER_NOT_AVAILABLE.code(), response.error().code());
@@ -236,14 +236,14 @@ class NetworkPartitionMetadataClientTest {
         Set<TopicPartition> partitions = new HashSet<>();
         partitions.add(tp);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
         assertEquals(1, futures.size());
         assertTrue(futures.containsKey(tp));
 
-        PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
         assertTrue(futures.get(tp).isDone() && !futures.get(tp).isCompletedExceptionally());
         assertNotNull(response);
         assertEquals(Errors.LEADER_NOT_AVAILABLE.code(), response.error().code());
@@ -278,14 +278,14 @@ class NetworkPartitionMetadataClientTest {
         Set<TopicPartition> partitions = new HashSet<>();
         partitions.add(tp);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
         assertEquals(1, futures.size());
         assertTrue(futures.containsKey(tp));
 
-        PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
         assertTrue(futures.get(tp).isDone() && !futures.get(tp).isCompletedExceptionally());
         assertNotNull(response);
         assertEquals(Errors.UNKNOWN_SERVER_ERROR.code(), response.error().code());
@@ -294,8 +294,8 @@ class NetworkPartitionMetadataClientTest {
     @Test
     public void testListLatestOffsetsNullResponse() throws ExecutionException, InterruptedException {
         TopicPartition tp = new TopicPartition(TOPIC, PARTITION);
-        CompletableFuture<PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures = Map.of(
+        CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures = Map.of(
             tp,
             partitionFuture);
         networkPartitionMetadataClient = NetworkPartitionMetadataClientBuilder.builder().build();
@@ -308,7 +308,7 @@ class NetworkPartitionMetadataClientTest {
         // Pass null as clientResponse.
         networkPartitionMetadataClient.handleResponse(pendingReqeust, null);
         assertTrue(partitionFuture.isDone() && !partitionFuture.isCompletedExceptionally());
-        PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
         assertEquals(-1, response.offset());
         assertEquals(Errors.UNKNOWN_SERVER_ERROR.code(), response.error().code());
     }
@@ -316,8 +316,8 @@ class NetworkPartitionMetadataClientTest {
     @Test
     public void testListLatestOffsetsAuthenticationError() throws ExecutionException, InterruptedException {
         TopicPartition tp = new TopicPartition(TOPIC, PARTITION);
-        CompletableFuture<PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures = Map.of(
+        CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures = Map.of(
             tp,
             partitionFuture);
         AuthenticationException authenticationException = new AuthenticationException("Test authentication exception");
@@ -333,7 +333,7 @@ class NetworkPartitionMetadataClientTest {
             builder);
         networkPartitionMetadataClient.handleResponse(pendingReqeust, clientResponse);
         assertTrue(partitionFuture.isDone() && !partitionFuture.isCompletedExceptionally());
-        PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
         assertEquals(-1, response.offset());
         assertEquals(Errors.UNKNOWN_SERVER_ERROR.code(), response.error().code());
     }
@@ -341,8 +341,8 @@ class NetworkPartitionMetadataClientTest {
     @Test
     public void testListLatestOffsetsVersionMismatch() throws ExecutionException, InterruptedException {
         TopicPartition tp = new TopicPartition(TOPIC, PARTITION);
-        CompletableFuture<PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures = Map.of(
+        CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures = Map.of(
             tp,
             partitionFuture);
         UnsupportedVersionException unsupportedVersionException = new UnsupportedVersionException("Test unsupportedVersionException exception");
@@ -359,7 +359,7 @@ class NetworkPartitionMetadataClientTest {
             builder);
         networkPartitionMetadataClient.handleResponse(pendingReqeust, clientResponse);
         assertTrue(partitionFuture.isDone() && !partitionFuture.isCompletedExceptionally());
-        PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
         assertEquals(-1, response.offset());
         assertEquals(Errors.UNKNOWN_SERVER_ERROR.code(), response.error().code());
     }
@@ -418,7 +418,7 @@ class NetworkPartitionMetadataClientTest {
         partitions.add(tp1);
         partitions.add(tp2);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
@@ -426,13 +426,13 @@ class NetworkPartitionMetadataClientTest {
         assertTrue(futures.containsKey(tp1));
         assertTrue(futures.containsKey(tp2));
 
-        PartitionMetadataClient.OffsetResponse response1 = futures.get(tp1).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response1 = futures.get(tp1).get();
         assertTrue(futures.get(tp1).isDone() && !futures.get(tp1).isCompletedExceptionally());
         assertNotNull(response1);
         assertEquals(Errors.NONE.code(), response1.error().code());
         assertEquals(expectedOffset1, response1.offset());
 
-        PartitionMetadataClient.OffsetResponse response2 = futures.get(tp2).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response2 = futures.get(tp2).get();
         assertTrue(futures.get(tp2).isDone() && !futures.get(tp2).isCompletedExceptionally());
         assertNotNull(response2);
         assertEquals(Errors.NONE.code(), response2.error().code());
@@ -516,7 +516,7 @@ class NetworkPartitionMetadataClientTest {
         partitions.add(tp1);
         partitions.add(tp2);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
@@ -524,13 +524,13 @@ class NetworkPartitionMetadataClientTest {
         assertTrue(futures.containsKey(tp1));
         assertTrue(futures.containsKey(tp2));
 
-        PartitionMetadataClient.OffsetResponse response1 = futures.get(tp1).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response1 = futures.get(tp1).get();
         assertTrue(futures.get(tp1).isDone() && !futures.get(tp1).isCompletedExceptionally());
         assertNotNull(response1);
         assertEquals(Errors.NONE.code(), response1.error().code());
         assertEquals(expectedOffset1, response1.offset());
 
-        PartitionMetadataClient.OffsetResponse response2 = futures.get(tp2).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response2 = futures.get(tp2).get();
         assertTrue(futures.get(tp2).isDone() && !futures.get(tp2).isCompletedExceptionally());
         assertNotNull(response2);
         assertEquals(Errors.NONE.code(), response2.error().code());
@@ -595,7 +595,7 @@ class NetworkPartitionMetadataClientTest {
         partitions.add(tp1);
         partitions.add(tp2);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
@@ -603,13 +603,13 @@ class NetworkPartitionMetadataClientTest {
         assertTrue(futures.containsKey(tp1));
         assertTrue(futures.containsKey(tp2));
 
-        PartitionMetadataClient.OffsetResponse response1 = futures.get(tp1).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response1 = futures.get(tp1).get();
         assertTrue(futures.get(tp1).isDone() && !futures.get(tp1).isCompletedExceptionally());
         assertNotNull(response1);
         assertEquals(Errors.NONE.code(), response1.error().code());
         assertEquals(expectedOffset1, response1.offset());
 
-        PartitionMetadataClient.OffsetResponse response2 = futures.get(tp2).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response2 = futures.get(tp2).get();
         assertTrue(futures.get(tp2).isDone() && !futures.get(tp2).isCompletedExceptionally());
         assertNotNull(response2);
         assertEquals(Errors.NONE.code(), response2.error().code());
@@ -620,7 +620,7 @@ class NetworkPartitionMetadataClientTest {
     public void testListLatestOffsetsNullPartitions() {
         networkPartitionMetadataClient = NetworkPartitionMetadataClientBuilder.builder().build();
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(null);
 
         assertNotNull(futures);
@@ -633,7 +633,7 @@ class NetworkPartitionMetadataClientTest {
 
         Set<TopicPartition> partitions = new HashSet<>();
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
@@ -680,14 +680,14 @@ class NetworkPartitionMetadataClientTest {
         Set<TopicPartition> partitions = new HashSet<>();
         partitions.add(tp);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
         assertEquals(1, futures.size());
         assertTrue(futures.containsKey(tp));
 
-        PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
         assertTrue(futures.get(tp).isDone() && !futures.get(tp).isCompletedExceptionally());
         assertNotNull(response);
         assertEquals(Errors.UNKNOWN_TOPIC_OR_PARTITION.code(), response.error().code());
@@ -729,14 +729,14 @@ class NetworkPartitionMetadataClientTest {
         Set<TopicPartition> partitions = new HashSet<>();
         partitions.add(tp);
 
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures =
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures =
             networkPartitionMetadataClient.listLatestOffsets(partitions);
 
         assertNotNull(futures);
         assertEquals(1, futures.size());
         assertTrue(futures.containsKey(tp));
 
-        PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = futures.get(tp).get();
         assertTrue(futures.get(tp).isDone() && !futures.get(tp).isCompletedExceptionally());
         assertNotNull(response);
         assertEquals(Errors.UNKNOWN_TOPIC_OR_PARTITION.code(), response.error().code());
@@ -969,8 +969,8 @@ class NetworkPartitionMetadataClientTest {
     @Test
     public void testRetryOnDisconnect() {
         TopicPartition tp = new TopicPartition(TOPIC, PARTITION);
-        CompletableFuture<PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures = Map.of(
+        CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures = Map.of(
             tp,
             partitionFuture);
         MockTimer timer = new MockTimer(MOCK_TIME);
@@ -1014,8 +1014,8 @@ class NetworkPartitionMetadataClientTest {
     @Test
     public void testRetryOnTimeout() {
         TopicPartition tp = new TopicPartition(TOPIC, PARTITION);
-        CompletableFuture<PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures = Map.of(
+        CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures = Map.of(
             tp,
             partitionFuture);
         MockTimer timer = new MockTimer(MOCK_TIME);
@@ -1060,8 +1060,8 @@ class NetworkPartitionMetadataClientTest {
     @Test
     public void testMaxRetryAttemptsExhaustedOnDisconnect() throws ExecutionException, InterruptedException {
         TopicPartition tp = new TopicPartition(TOPIC, PARTITION);
-        CompletableFuture<PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures = Map.of(
+        CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures = Map.of(
             tp,
             partitionFuture);
         MockTimer timer = new MockTimer(MOCK_TIME);
@@ -1107,7 +1107,7 @@ class NetworkPartitionMetadataClientTest {
         assertEquals(0, timer.size());
         // Verify that future is completed with error
         assertTrue(partitionFuture.isDone());
-        PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
         assertEquals(-1, response.offset());
         assertEquals(Errors.NETWORK_EXCEPTION.code(), response.error().code());
     }
@@ -1115,8 +1115,8 @@ class NetworkPartitionMetadataClientTest {
     @Test
     public void testMaxRetryAttemptsExhaustedOnTimeout() throws ExecutionException, InterruptedException {
         TopicPartition tp = new TopicPartition(TOPIC, PARTITION);
-        CompletableFuture<PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
-        Map<TopicPartition, CompletableFuture<PartitionMetadataClient.OffsetResponse>> futures = Map.of(
+        CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse> partitionFuture = new CompletableFuture<>();
+        Map<TopicPartition, CompletableFuture<org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse>> futures = Map.of(
             tp,
             partitionFuture);
         MockTimer timer = new MockTimer(MOCK_TIME);
@@ -1163,7 +1163,7 @@ class NetworkPartitionMetadataClientTest {
         assertEquals(0, timer.size(), "Timer should not have an entry when max retries are exhausted");
         // Verify that future is completed with error
         assertTrue(partitionFuture.isDone(), "Future should be completed when max retries are exhausted");
-        PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
+        org.apache.kafka.server.util.PartitionMetadataClient.OffsetResponse response = partitionFuture.get();
         assertEquals(-1, response.offset());
         assertEquals(Errors.REQUEST_TIMED_OUT.code(), response.error().code());
     }
